entity design

+++ src/main/java/com/hotelgroup/lakeSide_hotel/repository/BookedRoomRepository.java
@@ -3,5 +3,5 @@ package com.hotelgroup.lakeSide_hotel.repository;
 import com.hotelgroup.lakeSide_hotel.model.BookedRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookedRoomRepository extends JpaRepository<BookedRoom,Integer> {
+public interface BookedRoomRepository extends JpaRepository<BookedRoom,Long> {
 }
